Refactor: Admin page 분리.
